Ruby != Ruby on Rails
Ruby is a programming language and one that is needed to compile Sass (still not 100% true: there are multiple ports to other languages, most importantly [libsass](https://github.com/sass/libsass) which is a C/C++ implementation which makes it easier to port to other languages, i.e. Node via [node-sass](https://www.npmjs.org/package/node-sass) (i.e. JavaScript)).

Ruby on Rails is a web framework written in Ruby that is not needed to compile Sass.

As another sidenote, On Sass' website, they stylize it Sass and not SASS.